+++ src/index.html
@@ -13,8 +13,10 @@
     <!-- partials -->
     <include src="./partials/header.html"></include>
     <include src="./partials/categories.html"></include>
+    <include src="./partials/support.html"></include>
     <include src="./partials/book-card.html"></include>
     <include src="./partials/authorization-modal.html"></include>
+    <include src="./partials/to-top.html"></include>
     <!-- scrypts  -->
     <script type="module" src="./index.js"></script>
   </body>
