+++ src/js/firebase/firebase-user.js
@@ -1,6 +1,4 @@
-import { initializeApp } from 'firebase/app';
-import { getFirestore, collection, getDocs } from 'firebase/firestore/lite';
-import { getStorage } from 'firebase/storage';
+import { app } from './firebace-config';
 import {
   getAuth,
   createUserWithEmailAndPassword,
@@ -9,6 +7,9 @@ import {
   signOut,
 } from 'firebase/auth';
 import Notiflix from 'notiflix';
+
+// Initialize Firebase Authentication and get a reference to the service
+const auth = getAuth(app);
 
 const refs = {
   autirizationFormEl: document.querySelector('.authorization__form'),
@@ -48,27 +49,6 @@ refs.autirizationFormEl.addEventListener('submit', handelRegistrUser);
 refs.autirizationFormEl.addEventListener('submit', handelSignInUserAccount);
 refs.logOutBtn.addEventListener('click', handelLogOutUserAccount);
 
-const firebaseConfig = {
-  apiKey: 'AIzaSyDfT-rZpY0OALd7KjMrkrFQZlQZKUmMqYA',
-  authDomain: 'book-159d3.firebaseapp.com',
-  databaseURL: 'https://book-159d3-default-rtdb.firebaseio.com',
-  projectId: 'book-159d3',
-  storageBucket: 'book-159d3.appspot.com',
-  storageBucket: 'book-159d3.appspot.com',
-  messagingSenderId: '35676359423',
-  appId: '1:35676359423:web:0fe005890081fa3d52daca',
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const db = getFirestore(app);
-
-// Initialize Cloud Storage and get a reference to the service
-const storage = getStorage(app);
-
-// Initialize Firebase Authentication and get a reference to the service
-const auth = getAuth(app);
-
 function handelRegistrUser(evt) {
   evt.preventDefault();
 
@@ -99,14 +79,13 @@ function handelRegistrUser(evt) {
             'A user with this email address is already registered'
           );
         }
-        console.log(error.message);
       });
   }
 }
 
 function handelSignInUserAccount(evt) {
   evt.preventDefault();
-  console.log(evt.target);
+
   const {
     elements: { name, email, password },
   } = evt.currentTarget;
